agent-server/requirements.txt
# --- Core ---
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-frontmatter
jinja2

# --- LangChain / LangGraph ---
langchain
langgraph
langchain-openai
langsmith                    # test용

# --- firebase ---
firebase-admin
google-api-core
langchain-google-firestore

## ======== RAG ======== ##
# --- Vector DB: Pinecone ---
pinecone                # Pinecone 최신 SDK(v5)
langchain-pinecone             # LangChain ↔ Pinecone 연결 어댑터
langchain_huggingface
langchain_community           # LangChain 커뮤니티 모듈들(여기서 PineconeVectorStore 사용)

# --- Embedding / 문서처리 ---
sentence-transformers          # (Hugging Face 임베딩용)
pypdf                          # PDF 로더
faiss-cpu                      # 과거 대화 내역 임베딩 검색용 Vector DB

# --- 기타 유틸 ---
tqdm
PyYAML                         # protocol 파일 형식 YAML 임포트(app.utils.protocol_loader.py)